Address some N+1 queries in `SpeciesSerializer`
From looking at the [Skylight dashboard][1], we found some N+1/
repeated queries on the `Api::V1::SpeciesController#show` endpoint.

In the `SpeciesSerializer`, there are a few `has_many` associations
that needs to be serialized. Some of these associations then uses
serializers that themselves have other `has_many` associations that
needs to be serialized. This resulted in the classic [N+1 queries][2]
problem.

The usual way to address is to add an `includes` to the query so
Rails knows to eager load all the nested associations with a single
query. However, in this case, the query is built by `Panko` and there
does not appear to be a way to tell `Panko` to do this.

As a workaround, we turned the `has_many` associations to a normal
attribute, in which we manually build the query with the preload
instruction and serialize them using `ArraySerializer`.

The tests appear to be broken even before this commit, so we are
unable to run them against these changes. However, we visited the
API endpoint locally and saw no changes to the JSON output, other
than a few ordering differences.

[1]: https://oss.skylight.io/app/applications/nz7MAOv6K6ra/recent/6h/endpoints/Api::V1::SpeciesController%23show?responseType=json
[2]: https://www.skylight.io/support/performance-tips#repeated-queries

Co-authored-by: Peter Wagenet <peter.wagenet@gmail.com>